Use Microsoft DI on demo
