+++ src/StarDust.CasparCG.net.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29306.81
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33213.308
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StarDust.CasparCg.net.Models", "StarDust.CasparCG.net.Models\StarDust.CasparCg.net.Models.csproj", "{70932588-C180-4B2E-933F-BB5ADEEB8B34}"
 EndProject
@@ -35,7 +35,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StartDust.CasparCG.net.OSC.
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demo WPF", "Demo WPF", "{7B9EF232-F891-4464-8119-6CBD94BC2464}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StarDust.CasparCG.net.Demo.WPF", "Demo\Demo WPF\StarDusrt.CasparCG.net.Demo.WPF\StarDust.CasparCG.net.Demo.WPF.csproj", "{938A7B19-A906-48C8-847B-5BF590CFDEEF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StarDust.CasparCG.net.Demo.WPF", "Demo\Demo WPF\StarDusrt.CasparCG.net.Demo.WPF\StarDust.CasparCG.net.Demo.WPF.csproj", "{938A7B19-A906-48C8-847B-5BF590CFDEEF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StarDust.CasparCG.net.Microsoft.DependencyInjection", "StarDust.CasparCG.net.Microsoft.DependencyInjection\StarDust.CasparCG.net.Microsoft.DependencyInjection.csproj", "{6128B0BB-4A38-4664-ACAB-7DB09FAFF56D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -91,18 +93,22 @@ Global
 		{938A7B19-A906-48C8-847B-5BF590CFDEEF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{938A7B19-A906-48C8-847B-5BF590CFDEEF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{938A7B19-A906-48C8-847B-5BF590CFDEEF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6128B0BB-4A38-4664-ACAB-7DB09FAFF56D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6128B0BB-4A38-4664-ACAB-7DB09FAFF56D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6128B0BB-4A38-4664-ACAB-7DB09FAFF56D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6128B0BB-4A38-4664-ACAB-7DB09FAFF56D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{7B9EF232-F891-4464-8119-6CBD94BC2464} = {0FFA6423-92B3-4CE2-ACA1-1358FB3C7355}
 		{E4E01B5D-63FD-48DB-96D8-BF2B73CEDE79} = {0FFA6423-92B3-4CE2-ACA1-1358FB3C7355}
 		{C155102F-AD87-4DF4-8023-9CB2C8A0CDA9} = {E4E01B5D-63FD-48DB-96D8-BF2B73CEDE79}
 		{F16A5C88-2058-4597-8418-1F67244A7E63} = {D20E0CA0-35E6-4512-9DC2-9519B2C7DD36}
 		{2CE679AD-F37F-4E21-BD19-4057F9E34B06} = {0FFA6423-92B3-4CE2-ACA1-1358FB3C7355}
 		{0D1B3736-FB7D-45C8-AF25-E09CE0796011} = {2CE679AD-F37F-4E21-BD19-4057F9E34B06}
 		{18EFED87-58BA-4018-BC0A-BB385B32938C} = {D20E0CA0-35E6-4512-9DC2-9519B2C7DD36}
+		{7B9EF232-F891-4464-8119-6CBD94BC2464} = {0FFA6423-92B3-4CE2-ACA1-1358FB3C7355}
 		{938A7B19-A906-48C8-847B-5BF590CFDEEF} = {7B9EF232-F891-4464-8119-6CBD94BC2464}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
